// types/financial-metrics.ts
export interface FinancialMetrics {
  peRatio?: number;
  eps?: number;

  beta?: number;

  revenueGrowth?: number;

  marketCap?: number;

  week52High?: number;
  week52Low?: number;

  sharesOutstanding?: number;

  grossMargin?: number;
  operatingMargin?: number;
  netMargin?: number;

  roe?: number;
  roa?: number;

  currentRatio?: number;

  debtToEquity?: number;

  freeCashFlowPerShare?: number;
}